+++ src/types/GTableHeader.ts
@@ -8,4 +8,5 @@ export interface GTableHeader {
   checkboxHeader?: boolean; // optional, set to false when checkbox type is set and the toggle all checkbox in header should not be displayed
   isChecked?: (arg: any) => boolean; // callback function to get the checked status if item is a checkbox
   render?: (arg: any) => string; // callback function to get the content to be rendered of the item
+  sortable?: boolean; // optional, default false. If true and field is set (and type is not "checkbox"/"expandable"), the column header is clickable for sorting
 }